Moved classes for referral cancellation from /server/routes/probationPractitionerReferrals to /server/routes/referral/cancellation
3 directories were created within the cancellation directory:

./checkAnswers
./confirmation
./reason

Moved referral cancellation specific logic into its own controller in  /server/routes/referral/cancellation
