+++ server/routes/probationPractitionerReferrals/probationPractitionerReferralsController.ts
@@ -11,16 +11,9 @@ import DashboardPresenter, { PPDashboardType } from './dashboardPresenter'
 import FindStartView from './findStartView'
 import SubmittedFeedbackPresenter from '../shared/appointment/feedback/submittedFeedbackPresenter'
 import SubmittedFeedbackView from '../shared/appointment/feedback/submittedFeedbackView'
-import ReferralCancellationReasonPresenter from './referralCancellationReasonPresenter'
-import ReferralCancellationReasonView from './referralCancellationReasonView'
 import EndOfServiceReportPresenter from '../shared/endOfServiceReport/endOfServiceReportPresenter'
 import EndOfServiceReportView from '../shared/endOfServiceReport/endOfServiceReportView'
-import ReferralCancellationCheckAnswersPresenter from './referralCancellationCheckAnswersPresenter'
-import ReferralCancellationCheckAnswersView from './referralCancellationCheckAnswersView'
 import { FormValidationError } from '../../utils/formValidationError'
-import ReferralCancellationReasonForm from './referralCancellationReasonForm'
-import ReferralCancellationConfirmationView from './referralCancellationConfirmationView'
-import ReferralCancellationConfirmationPresenter from './referralCancellationConfirmationPresenter'
 import ControllerUtils from '../../utils/controllerUtils'
 import ShowReferralPresenter from '../shared/showReferralPresenter'
 import ShowReferralView from '../shared/showReferralView'
@@ -35,7 +28,6 @@ import SupplierAssessmentAppointmentPresenter from '../shared/supplierAssessment
 import SupplierAssessmentAppointmentView from '../shared/supplierAssessmentAppointmentView'
 import FileUtils from '../../utils/fileUtils'
 import DraftsService from '../../services/draftsService'
-import DraftCancellationData from './draftCancellationData'
 import config from '../../config'
 import AppointmentSummary from '../appointments/appointmentSummary'
 import DeliusOfficeLocationFilter from '../../services/deliusOfficeLocationFilter'
@@ -314,141 +306,6 @@ export default class ProbationPractitionerReferralsController {
     const view = new SubmittedFeedbackView(presenter)
 
     return ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async startCancellation(req: Request, res: Response): Promise<void> {
-    const draftCancellation = await this.draftsService.createDraft<DraftCancellationData>(
-      'cancellation',
-      {
-        cancellationReason: null,
-        cancellationComments: null,
-      },
-      { userId: res.locals.user.userId }
-    )
-    res.redirect(`/probation-practitioner/referrals/${req.params.id}/cancellation/${draftCancellation.id}/reason`)
-  }
-
-  private async fetchDraftCancellationOrRenderMessage(req: Request, res: Response) {
-    return ControllerUtils.fetchDraftOrRenderMessage<DraftCancellationData>(req, res, this.draftsService, {
-      idParamName: 'draftCancellationId',
-      notFoundUserMessage:
-        'Too much time has passed since you started cancelling this referral. Your answers have not been saved, and you will need to start again.',
-      typeName: 'cancellation',
-    })
-  }
-
-  async editCancellationReason(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    const fetchResult = await this.fetchDraftCancellationOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-    const draftCancellation = fetchResult.draft
-
-    const data = await new ReferralCancellationReasonForm(req).data()
-
-    let formError: FormValidationError | null = null
-
-    if (req.method === 'POST') {
-      if (!data.error) {
-        await this.draftsService.updateDraft<DraftCancellationData>(draftCancellation.id, data.paramsForUpdate, {
-          userId: res.locals.user.userId,
-        })
-
-        res.redirect(
-          `/probation-practitioner/referrals/${referralId}/cancellation/${draftCancellation.id}/check-your-answers`
-        )
-        return
-      }
-
-      res.status(400)
-      formError = data.error
-    }
-
-    const sentReferral = await this.interventionsService.getSentReferral(accessToken, referralId)
-    const intervention = await this.interventionsService.getIntervention(
-      accessToken,
-      sentReferral.referral.interventionId
-    )
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(sentReferral.referral.serviceUser.crn)
-    const cancellationReasons = await this.interventionsService.getReferralCancellationReasons(accessToken)
-
-    const presenter = new ReferralCancellationReasonPresenter(
-      draftCancellation,
-      sentReferral,
-      intervention,
-      serviceUser,
-      cancellationReasons,
-      formError
-    )
-    const view = new ReferralCancellationReasonView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async cancellationCheckAnswers(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-
-    const fetchResult = await this.fetchDraftCancellationOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-    const draftCancellation = fetchResult.draft
-
-    const sentReferral = await this.interventionsService.getSentReferral(accessToken, req.params.id)
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(sentReferral.referral.serviceUser.crn)
-
-    const presenter = new ReferralCancellationCheckAnswersPresenter(req.params.id, draftCancellation.id)
-    const view = new ReferralCancellationCheckAnswersView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
-  }
-
-  async submitCancellation(req: Request, res: Response): Promise<void> {
-    const fetchResult = await this.fetchDraftCancellationOrRenderMessage(req, res)
-    if (fetchResult.rendered) {
-      return
-    }
-    const draftCancellation = fetchResult.draft
-
-    const { cancellationReason, cancellationComments } = draftCancellation.data
-
-    if (cancellationReason === null) {
-      throw new Error('Got unexpectedly null cancellationReason')
-    }
-    if (cancellationComments === null) {
-      throw new Error('Got unexpectedly null cancellationComments')
-    }
-
-    const { user } = res.locals
-    const { accessToken } = user.token
-    const referralId = req.params.id
-
-    await this.interventionsService.endReferral(accessToken, referralId, cancellationReason, cancellationComments)
-    await this.draftsService.deleteDraft(draftCancellation.id, { userId: res.locals.user.userId })
-
-    res.redirect(`/probation-practitioner/referrals/${referralId}/cancellation/confirmation`)
-  }
-
-  async showCancellationConfirmationPage(req: Request, res: Response): Promise<void> {
-    const { user } = res.locals
-    const { accessToken } = user.token
-
-    const sentReferral = await this.interventionsService.getSentReferral(accessToken, req.params.id)
-    const intervention = await this.interventionsService.getIntervention(
-      accessToken,
-      sentReferral.referral.interventionId
-    )
-    const serviceUser = await this.communityApiService.getServiceUserByCRN(sentReferral.referral.serviceUser.crn)
-
-    const presenter = new ReferralCancellationConfirmationPresenter(sentReferral, intervention)
-    const view = new ReferralCancellationConfirmationView(presenter)
-
-    ControllerUtils.renderWithLayout(res, view, serviceUser)
   }
 
   async viewEndOfServiceReport(req: Request, res: Response): Promise<void> {
